Add Polygon entries in coins
Fixes https://app.clickup.com/t/37308523/PRD-1453

+++ common/coin_support/coin_utils.h
@@ -274,7 +274,7 @@ void get_address_node(const txn_metadata *txn_metadata_ptr, const int16_t index,
  *
  * @note
  */
-const char *get_coin_name(uint32_t coin_index, uint8_t chain_id);
+const char *get_coin_name(uint32_t coin_index, uint32_t chain_id);
 
 /**
  * @brief Get the coin symbol for the passed coin index and chain id
@@ -291,7 +291,7 @@ const char *get_coin_name(uint32_t coin_index, uint8_t chain_id);
  *
  * @note
  */
-const char *get_coin_symbol(int coin_index, uint8_t chain_id);
+const char *get_coin_symbol(uint32_t coin_index, uint32_t chain_id);
 
 /**
  * @brief Get the version address and public key for segwit and non segwit coins.
